exportador/requirements.txt
flask==3.0.3
python-dotenv==1.0.1
elasticsearch==8.15.1
pandas==2.2.3
xlsxwriter==3.2.0
